Added the right images for cards Dream Home and 90% financing
Those cards were coming up with no image on the webpage and the AI could not fix the problem. 

Earlier, when I was editing the site in Lovableai, it either did not render the images on the right card or loaded nothing at all in place of some images. 

A couple of times when Lovable did not put any image in some cards, I ended up loading the same images multiple times. 

It seems that when Lovable uploads and an image it automatically rename it.--but it keeps a reference to the original image name. 

But when the ai makes an error during the webpage rendering, for example, it does not render the image that was supposed to go on "Card A" does not render and "Card A" appears with its associated text but not its associated, image.  The developer not know why the image did not render, reloads the image that was supposed be there. Lovable will put the generic name for example 'Palm2.png' in the code but will but recognizing that name, will point to the "earlier renamed image in its "lovable_uploads" folder, and create a mismatch issue it can not easily fix on its own. It will it made the changes, but you will keep getting a card still with no images.  

In my case, when one of  the  images I uploaded "Palm2" did not post on the card, I uploaded it 2 more times, and it still did not render

In my case, the cimage: "/src/data folder was showing 
 image: "/lovable-uploads/Palm2.png"
but in the "lovable_uploads" the filename was written as:
ff3f2c8a-caa1-4efe-a171-403f8a3fc150.png

The same for 
 blackDeal.png    pointing to    
3482fbb8-15a8-4f60-b2b6-ca36f7ff9730.png  

beachfront  pointing to   
a9b17377-c565-45d2-98c3-326004ae6ce8.png   

To solve the problem, I edited the line  image: "/src/data  blogPost.tj  so that they match the ones in the 
 image: "/lovable-uploads/" folder

 image: "/src/data  blogPost.ts ff3f2c8a-caa1-4efe-a171-403f8a3fc150.png

image: "/lovable-uploads/",ff3f2c8a-caa1-4efe-a171-403f8a3fc150.png